requirements.txt
ftfy==6.1.1
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
regex==2022.6.2
scipy==1.8.1
setuptools==68.0.0
# setuptools.egg==info
tensorboardX==2.6
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.65.0
matplotlib